+++ server/agent-controller/src/main/java/org/bithon/server/agent/controller/api/AgentControllerApi.java
@@ -35,8 +35,8 @@ import org.bithon.server.agent.controller.service.AgentControllerServer;
 import org.bithon.server.agent.controller.service.AgentSettingLoader;
 import org.bithon.server.commons.exception.ErrorResponse;
 import org.bithon.server.discovery.declaration.controller.IAgentControllerApi;
-import org.bithon.server.web.service.security.jwt.JwtConfig;
-import org.bithon.server.web.service.security.jwt.JwtTokenComponent;
+import org.bithon.server.web.security.jwt.JwtConfig;
+import org.bithon.server.web.security.jwt.JwtTokenComponent;
 import org.bithon.shaded.com.google.protobuf.CodedInputStream;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.http.HttpStatus;
