+++ server/web-app/src/main/java/org/bithon/server/webapp/security/oauth2/OAuth2AuthSuccessHandler.java
@@ -19,8 +19,8 @@ package org.bithon.server.webapp.security.oauth2;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.bithon.component.commons.utils.StringUtils;
-import org.bithon.server.web.service.security.cookie.CookieHelper;
-import org.bithon.server.web.service.security.jwt.JwtTokenComponent;
+import org.bithon.server.web.security.cookie.CookieHelper;
+import org.bithon.server.web.security.jwt.JwtTokenComponent;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.oauth2.core.user.OAuth2UserAuthority;
