+++ server/web-app/src/main/java/org/bithon/server/webapp/security/oauth2/OAuth2AutoConfiguration.java
@@ -21,9 +21,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.bithon.component.commons.utils.CollectionUtils;
 import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.storage.InvalidConfigurationException;
-import org.bithon.server.web.service.security.IHttpSecurityCustomizer;
-import org.bithon.server.web.service.security.jwt.JwtConfig;
-import org.bithon.server.web.service.security.jwt.JwtTokenComponent;
+import org.bithon.server.web.security.IHttpSecurityCustomizer;
+import org.bithon.server.web.security.jwt.JwtConfig;
+import org.bithon.server.web.security.jwt.JwtTokenComponent;
 import org.springframework.boot.autoconfigure.security.oauth2.client.OAuth2ClientProperties;
 import org.springframework.boot.autoconfigure.security.oauth2.client.OAuth2ClientPropertiesMapper;
 import org.springframework.boot.autoconfigure.web.ServerProperties;
