+++ server/web-app/src/main/java/org/bithon/server/webapp/security/oauth2/OAuth2LoginAuthenticationEntryPoint.java
@@ -20,7 +20,7 @@ import groovy.util.logging.Slf4j;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.bithon.component.commons.utils.StringUtils;
-import org.bithon.server.web.service.security.cookie.CookieHelper;
+import org.bithon.server.web.security.cookie.CookieHelper;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
