+++ server/web-security/pom.xml
@@ -9,9 +9,9 @@
   </parent>
 
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>server-web-service-security</artifactId>
-  <name>server-web-service-security</name>
-  <description>Bithon Web Service Security</description>
+  <artifactId>server-web-security</artifactId>
+  <name>server-web-security</name>
+  <description>Bithon Web Security</description>
 
   <dependencies>
     <dependency>
@@ -33,10 +33,6 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-security</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-autoconfigure</artifactId>
     </dependency>
 
     <dependency>
