+++ server/web-security/src/main/java/org/bithon/server/web/security/api/SecurityApi.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.security;
+package org.bithon.server.web.security.api;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -26,8 +26,8 @@ import org.bithon.component.commons.utils.HumanReadableDuration;
 import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.commons.time.TimeSpan;
 import org.bithon.server.commons.utils.HumanReadableDurationConstraint;
-import org.bithon.server.web.service.security.jwt.JwtConfig;
-import org.bithon.server.web.service.security.jwt.JwtTokenComponent;
+import org.bithon.server.web.security.jwt.JwtConfig;
+import org.bithon.server.web.security.jwt.JwtTokenComponent;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
