+++ server/web-security/src/main/java/org/bithon/server/web/security/filter/JwtAuthenticationFilter.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.security.filter;
+package org.bithon.server.web.security.filter;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -25,8 +25,8 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.bithon.component.commons.utils.StringUtils;
-import org.bithon.server.web.service.security.cookie.CookieHelper;
-import org.bithon.server.web.service.security.jwt.JwtTokenComponent;
+import org.bithon.server.web.security.cookie.CookieHelper;
+import org.bithon.server.web.security.jwt.JwtTokenComponent;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
