+++ server/web-security/src/main/java/org/bithon/server/web/security/jwt/JwtTokenComponent.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.security.jwt;
+package org.bithon.server.web.security.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
